nova: handle main thread inside context area
Fixes #540

+++ base/include/base/crt0.h
@@ -19,17 +19,17 @@
  ** Program image exported symbols **
  ************************************/
 
-extern unsigned _prog_img_beg;  /* begin of program image (link address) */
-extern unsigned _prog_img_end;  /* end of program image */
+extern unsigned long _prog_img_beg;  /* begin of program image (link address) */
+extern unsigned long _prog_img_end;  /* end of program image */
 
 extern void (*_ctors_start)();  /* begin of constructor table */
 extern void (*_ctors_end)();    /* end of constructor table */
 extern void (*_dtors_start)();  /* begin of destructor table */
 extern void (*_dtors_end)();    /* end of destructor table */
 
-extern unsigned _start;         /* program entry point */
-extern unsigned _stack_low;     /* lower bound of intial stack */
-extern unsigned _stack_high;    /* upper bound of intial stack */
+extern unsigned long _start;         /* program entry point */
+extern unsigned long _stack_low;     /* lower bound of intial stack */
+extern unsigned long _stack_high;    /* upper bound of intial stack */
 
 
 /***************************************************
